Add a general transaction object for all products i.e. collection, disbursement, remittance
